+++ packages/camera/camera_web/example/integration_test/camera_bitrate_test.dart
@@ -93,9 +93,7 @@ void main() {
         cameraService.getMediaStreamForOptions(options, cameraId: cameraId),
       ).thenAnswer((_) async => canvasElement.captureStream());
 
-      when(
-        () => cameraService.hasPropertyOffScreenCanvas(),
-      ).thenAnswer((_) => true);
+      when(cameraService.hasPropertyOffScreenCanvas()).thenAnswer((_) => true);
 
       final Camera camera = Camera(
         textureId: cameraId,