+++ packages/camera/camera_web/CHANGELOG.md
@@ -1,6 +1,5 @@
-## 0.3.5+1
+## NEXT
 
-* Fixes flutter/flutter#148013: setDescriptionWhileRecording with android camerax.
 * Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 
 ## 0.3.5