+++ packages/camera/camera_web/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.3.6
 
-- Supporting camera image stream on web.
+- Adds support for camera image stream on web.
 - Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 
 ## 0.3.5